requirements.txt
anyio==3.6.2
atomicwrites==1.4.1
attrs==19.3.0
beautifulsoup4==4.11.1
certifi==2019.9.11
chardet==3.0.4
charset-normalizer==2.1.1
colorama==0.4.5
cycler==0.11.0
Django==2.2.6
django-debug-toolbar==2.2
django-embed-video==1.4.8
Faker==12.0.1
flake8==5.0.4
fonttools==4.37.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==2.8
importlib-metadata==1.5.0
kiwisolver==1.4.4
mccabe==0.7.0
mixer==7.1.2
more-itertools==8.2.0
numpy==1.21.6
packaging==20.1
Pillow==9.2.0
pluggy==0.13.1
py==1.8.1
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==2.4.6
pyTelegramBotAPI==4.7.0
pytest==5.3.5
pytest-django==3.8.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-telegram-bot==20.0
pytz==2019.3
requests==2.22.0
rfc3986==1.5.0
six==1.14.0
sniffio==1.3.0
sorl-thumbnail==12.6.3
soupsieve==2.3.2.post1
sqlparse==0.3.0
telebot==0.0.4
telegram==0.0.1
typing_extensions==4.3.0
urllib3==1.25.6
wcwidth==0.1.8
zipp==2.2.0
